+++ src/main/php/peer/Socket.php
@@ -305,16 +305,36 @@ class Socket
         return true;
     }
 
+    private $inputStream;
+
     /**
      * returns input stream to read from socket
      *
      * @return InputStream
      * @since  2.0.0
      */
+    public function in()
+    {
+        if (null === $this->inputStream) {
+            $this->inputStream = new SocketInputStream($this);
+        }
+
+        return $this->inputStream;
+    }
+
+    /**
+     * returns input stream to read from socket
+     *
+     * @return InputStream
+     * @since  2.0.0
+     * @deprecated  since 4.0.0, use in() instead, will be removed with 5.0.0
+     */
     public function getInputStream()
     {
-        return new SocketInputStream($this);
+        return $this->in();
     }
+
+    private $out;
 
     /**
      * returns output stream to write to socket
@@ -322,8 +342,24 @@ class Socket
      * @return OutputStream
      * @since  2.0.0
      */
+    public function out()
+    {
+        if (null === $this->out) {
+            $this->out = new SocketOutputStream($this);
+        }
+
+        return $this->out;
+    }
+
+    /**
+     * returns output stream to write to socket
+     *
+     * @return OutputStream
+     * @since  2.0.0
+     * @deprecated  since 4.0.0, use out() instead, will be removed with 5.0.0
+     */
     public function getOutputStream()
     {
-        return new SocketOutputStream($this);
+        return $this->out();
     }
 }